refactor(home): split the folders grid into a hook and helpers
Move the drag-and-drop logic into useHomeDnd and the geometry helpers
into dndGeometry, leaving index.tsx as a thin view (426 -> 125 lines).
